Move resource usage to sidebar menu
